Fix lessThanZ3 bug
lessThanZ3 was actually computing less-than-or-equals. Combined with an
overzealous implementation of disjointZ3, this caused lud to fail to properly
short-circuit (checking the internal part didn't succeed).  By changing
disjointZ3 to only check that lower_bound + number_of_elements of one interval
is less than or equal to the lower_bound of the other interval, full
functionality is restored.