When short-circuiting, source should inherit certificates of dest
This commit fixes the issue described in #1930 where short circuiting without
also applying the certificates of the destination to the source can cause
runtime errors and memory corruption.  In addition to propagating the
certificates, we also need to check that the values used in the certificate are
actually in scope.

Fixes #1930