Carefully track interprocedural aliasing in core language (#1958)
This is a proper fix of #803 (see also discussion in #1949). The main
issue we address is that in the core language representation of a
function originally returning [](t,t), the two constituent arrays
cannot possibly alias.

This fix complicates the core language, but in a principled way. Each
return type of a function is now accompanied by two lists:

    * Which function parameters the result may alias.
    * Which other results the result may alias.

For simplicity these are integer lists, interpreted as indexes. This
is quite trivial to check in the core typechecker and to propagate in
alias analysis. The main challenge is to infer the lists based on the
source language types, as the source language is not so precise about
aliasing. I don't feel that my current solution is fully principled,
and I'm worried that I might have gotten it wrong (which will manifest
as compiler crashes - not as quietly invalid code generation).

There is also a remaining outstanding question in that we have no
conscious policy on whether a core language function is allowed to
return something that aliases a global variable. We do not permit this
in the source language. We currently do allow this in the core, but I
think the only case where it will currently happen is for entry points
that are non-functions in the source program.

Fixes #1949.