Replace AnySize with sentinel expression.
I'm not a big fan of magical in-band signalling, which this certainly
is, but AnySize is a hack and an implementation flaw, and I'd rather
keep the AST as emitted by the type checker simple, for the sake of
other AST consumers.

There are very few places we use the new "anySize" expression, and
hopefully we can gradually eliminate them all.