Change ixfun base when reshaping.
That means this is not really the "original" or "underlying" shape.
In practice we use this as a vehicle to do certain transformations,
mostly related to memory expansion.  I would rather get entirely rid
of the notion (as in #1981), but it turns out to be very difficult.
Another day maybe.