Tweak manifestation of Match results.
Now we only do a copy if we do not already have the desired form.

I think this could actually be simplified further.  We don't actually
need to do a copy (unless the memory space is wrong) - we could just
tweak the index function to remove the permutation, since that is the
only case where we actually need to do these copies!

Perhaps we even need to get rid of permutations from index functions.