Check that dst is unused when short-circuiting function args
When short-circuiting with a function argument as the source, we should verify
that the destination is never used in-between the short-circuiting point and the
start of the function, because the source is live in all that code.  Instead of
checking for memory overlap, we can simply check that the destination is unused,
because we already enforce that the index function of the destination must match
the index function of the function argument.

Fixes #2010