Add map short-circuits after processing body of map
Previously, we erroneously added the short-circuiting attempt on maps before
processing the body. This led to us to reject some otherwise sound
short-circuitings.

Combined with the fix in handling function arguments as short-circuiting
sources, this erroneous behaviour caused
e.g. tests/memory-block-merging/coalescing/map/map.fut to fail:

```futhark
-- ==
-- input { [1,2,3] }
-- output { [2,3,4] }
-- structure gpu-mem { Alloc 0 }
-- structure seq-mem { Alloc 1 }

let main [n] (xs: *[n]i32): *[n]i32 =
  map (+1) xs
```

Trying to short-circuit `xs` (the source) into the result of the map (the
destination), the problem was that the (+1) (and resulting kernel body) resulted
in a destination use that interfered with the source. Since the function
argument is essentially "written to" at the start of the function, this caused
map.fut to not short-circuit.

Incidentally, this commit also reduces the number of allocations needed in a
handful of other tests.