BinOpIntermediateArrays.
`makeLocalArrays` now returns `[BinOpIntermediateArrays]`, where each
`BinOpIntermediateArrays` is the interms for a single `SegBinOp`, since
it works better with `SegBinOpSlug`. The rest of the module, and in
particular `SegBinOpSlug`, has been modified accordingly. Also, have
begun work on modifying the main loop in `...stageOne` for the non-comm
case. Finally, a number of small refactorings and removal of dead code.