This hopefully improves SegRed sequentialization
This commit modifies and, ideally, improves GPU code generation for
non-segmented and large-segments segmented reductions with
non-commutative and primitive operators (see description in module
header).

There are still some TODO's left to attend to -- most importantly, we
have yet to determine for certain exactly how big of an improvement --
if any at all -- this change is.

Also:
* Adds `nextMul` to `Util.IntegralExp` to standardize size alignment.
* Adds `isPrimParam` and `getChunkSize` (the latter lifted from
  `SegScan.SinglePass`) to `GPU.Base`, as well as minor refactorings to
  the module.
* Changes `SegScan.SinglePass` to use `getChunkSize`, and some light
  refactoring.
* Adds `tests/soacs/scan-in-loop.fut`, which test proper resetting and
  reuse of global mem dynamic ID counter.
* Removes `comment` from `ImpGen`, since it seemed to be a duplicate of
  `sComment`.
* CUDA compiled binaries now properly return 0 on successful
  `--dump-cuda`, similar to the other GPU backends.