LSP: Remove open/close handlers.
We don't actually need this information, since we reload the file
whenever necessary (when we get a hover request etc). And Eglot sends
a *ton* of these when opening a new file (one for each .fut file
already open), which manifests as Emacs hanging when you open a .fut
file when you already have several of them open.