Query register capacity for segred and segscan codegen.
This is very tedious code, and required adding the notion of "kernel
constant expressions", as we have some expressions that _must_ be
constant at kernel compilation time (which is at program runtime). We
actually had this notion in the ImpCode representation, but now ImpGen
provides some manual control as well.