Rework kernel-level constants in GPU backends.
Instead of being substituted into the kernel itself as a macro
definition, these are now computed on the host and passed in with -D.

The main advantage of this is that we no longer have to ensure that
the definition is a "constant expression" in C terms (which is
required to be usable for array sizes and such).

It may also ease the eventual implementation of non-C-based kernel
code generation.