Bump maxNumOps to 20.
User programs usually do not have many operators, but flattening has a
habit of splitting e.g operators like (\ (a,b) (c,d) -> (a+c,b+d+))
into two operators. This is fine because it reduces memory usage, but
it does also bloat the size of the kernels. Maybe we should try to
control this a bit better (or maybe just be slightly smarter in how we
generate code for multi-operator reduction kernels with the same
operator multiple times).