Remove unExistentialiseMemory.
This rather complicated and dubiously sound simplification rule has
long been necessary as a workaround for deficiencies elsewhere in the
compiler. With the recent improvements to double buffering and memory
expansion (and the general simplification of the memory
representation), it may no longer be necessary. Performance impact of
the removal is unclear, but the test suite works.