First working matrix multiplication
The shared memory indices in the register tile accumulation step was
missing some crucial parts. It now seems to work for arbitrary sizes and
tile sizes for at least the basic MM case, but I should test a little
more before I begin to generalize to arbitrary contractions.

Also, improved names for some of the Exps/SubExps for better legibility
in the generated code.