Interchange segmap and outer sequential dim loop
Wrt. the register tile accumulation loop nest: because the outermost Tq
loop (Tq being the sequential dimension tile size) was placed *outside*
the tblock-wide segmap, the synchronization which *should* follow the
entire loop nest was placed at the bottom of the Tq loop, which gave an
extra factor Tq synchronizations. This was fixed by interchanging this
outer loop inside the segmap.