Ditch the SegVirt on copying from global to shared
The virtualization loop in the LMAD-like copies from global to shared
memory can (and was previously) easily and concisely be implemented
using SegVirt. However, SegVirt wraps the entire loop body in a
boundary guard of `(i < tile_size_flat)`, whereas we only want that
guard on the write to shmem.