Automatic shmem padding to avoid bank conflicts
Now automatically detects candidate arrays for shmem padding. A shmem
array is candidate for shmem padding *if* one of its initial dimensions
are indexed by the FVI of the tblock. This analysis is quite simple.
Then, padding is added at (host code) runtime only if the product of
sizes of the inner dimensions is even.

To add as little padding as necessary, shmem are now flat. To explain
why this was necessary, here's an example: say our shmem array has
dimensions [Ta][Tb][Tc][Td], and the second dimension (of size Tb) is
indexed by the fastest varying thread index. Then we need to pad the
innermost two dimensions if (Tc * Td) is even. The simple, but expensive
solution is to pad the array to size [Ta][Tb][Tc+1][Td+1]. However, this
can quickly explode in space if the number of inner dimensions that need
padding is large. Also, it requires runtime analysis to find out which
of Tc and Td is even, since we know only that at least one is even if Tc
* Td is even, but not which one -- if only one is even and we pad both,
  then we still get bank conflicts.

The chosen solution is to flatten the shmem dimensions and padding those
which need padding together. In the above example, this means flattening
the dimensions to [Ta * Tb * Tc * Td] and padding using [Ta * Tb * (Tc *
Td + 1)]. However, since the array can no longer be represented as a
regular 4D array, it must stay flattened, which makes for slightly more
tedious index calculations. Note that we obviously still need to check
parity (in this case, of Tc * Td) at runtime.