Optimize the MM case
Optimization specifically for the MM case: As it turned out, LMAD
copying adds excessive overhead in the MM case, possibly due to
reshaping the tblock. This replaces the flat LMAD copy loop with a loop
nest similar to that of BlkRegTiling specifically for the MM case.

Currently it is a copy+paste of the existing copyGlb2Shr function.
Should probably clean up the code if it turns out to be beneficial.