Futhark.Test: refactor a bit.
In particular, errors are now single T.Texts rather than lists. The
lists are an unfortunate design in Futhark.CLI.Test that we should not
allow to infect the rest of the system.

The main functionality change is that 'auto output' generation is now
done with server-mode executables. This means we have no more use of
non-server executables in the tooling.