Distinguish holes from variables during unification.
Fixes a bug that would lead to cycles in the
substitution where a hole would be unified with
(and substituted for) a variable and then this variable
would later be unified with a different variable,
as if it were a hole. The paper I followed considers
all variables to be holes, hence the bug here.