Fix subtle bug by discerning replacements from substitutions.
A Substitution is now a Replacement with an associated name source.
A Replacement is a mapping from VNames to things that are
to be replaced. Substitution is no longer a Monoid.

Before this, writing `sub (mkSub a b) x` would subtly rename
`x` using a blank name source because mempty for Substitution
created a blank name source. Now it is impossible to write this.