Use monadic definition of Fmt.
This moves around the uses of 'fmt' to be explicit in the printer
itself rather than in the combinators. I propose this as a more
explicit design.

The one exception is (<+/>), which I think is a symptom that something
is wrong.

As long as Fmt must be an instance of Located, then there will be
bugs.