MINIMALLY INVASIVE refinement annotations.
NOTE this commit probably introduces errors in the type checker.

This commit adds Lubin Bailly's refinement annotations, not caring
about their being checked, just that we have the annotations.
It is actually checked that the annotation is a function
from the annotated type to bool, however.

Any unnecessary changes that I deemed "too invasive" have been left out.
For example, none of the changes to sliceShape are included.