Lex `-- |>` as the beginning of a regular comment
I somewhat often find myself commenting lines beginning with a forward
pipe (e.g. when chaining pipe operators) and having to manually insert
an extra dummy character to avoid the "Doc comments can only precede
declarations" parser error. With this change a `DOC` token is only lexed
if the `|` is not *immediately* succeeded by a `>`.