Fix inc. ranges with start == end and negative step (2) (#2195)
This changes behavior of inclusive ranges for the case where
`start == end` and `start != second`, for compiled programs only.
In this case the result is now `[start]` rather than a runtime error.

Example: `1..0...1` now produces `[1]` rather than an "invalid range"
runtime error.

This makes behavior for compiled programs similar to that of the Futhark
interpreter (and e.g. Haskell ranges), for at least this particular
case (I have not looked for other discrepancies).