Fix bugs in range/equivalence environments.
Fixes a logical error in the check for valid candidates
when refining (adding and substituting) equivalences
into the environment.

Restricts ranges to be on variables, avoiding ranges of the form
  {} <= i <= Sum x[0:k]
  {} <= Sum x[0:k-1] <= i
from being added to the environment. (The "new" symbol
Sum x[0:k-1] (note k-1) is preferred to completing
the range on i when considering the range candidate.)
Now the range on i is completed instead:
  Sum x[0:k-1] <= i <= Sum x[0:k]