Fix bug when mapping over terms in a SoP.
Use of sopFromList is unsafe when mapping terms. I had a SoP whose
terms contained quantified expressions with unique quantifiers
(risky territory to begin with).
Renaming the quantifiers would _delete_ terms from a SoP due
to the way SoPs are represented as Maps with terms
as keys. This fixes that.
For example,
  renameM (Sum_j1 e1(j2) - Sum_j2 e1(j2))
resulted in
  ==> Sum_j3 e1(j2).
  (Keys were duplicate after renaming, I guess.)
When it should have resulted in
  ==> Sum_j3 e1(j3) - Sum_j3 e1(j3)
  ==> 0.