Fix bug causing cycles in range environment and support fancier bounds checking.
With this fix, ranges can now again be added on symbols
rather than just variable names.
(Commit 3a7a289 restricted ranges to variable
names to combat cycles and get desirable
ranges. It is described there.)

Ranges on symbols are needded to do bounds checking
on expressions like

  if 0 < e and e <= length xs then xs[e - 1] else 0

where e can be any expression (and hence SoP symbol).

The bug fix:
The generic function for computing the transitive
closure of symbols through the range environment
is now extendendable by defining how symbols
"relate" to eachother in the ranges.
This lets me recreate the behaviour of Cosmin's
symbol specific transClosInRangesSym.
(This function is kept around for now because it
detects and reports cycles nicely.)

Finally, this meant a change in what symbols
are picked for bounding in the range envirnoment
and the Query module was changed to reflect this.
(To get the desired ranges from index functions.)