Fix #2209.
This required a rework of the handling of size expressions in the
interpreter. The problem is that Futhark's type rules are based on
sizes being non-interpreted expressions, as we need to be able to
match symbolic expressions to fish out concrete values for size
parameters. However sometimes we also need to evaluate those
expressions to integers, and in those cases we better have the right
environment at hand for doing so. The interpreter did not properly do
so, in some cases.

The problem was related to the module system, which is properly
defined in terms of name substitution (and this is how the compiler
itself implements it). However, the interpreter is based on a
philosophy of not doing any transformation of the program prior to
evaluation (in order to "keep us honest" about the semantics), which
means being careful about environments. Now all size expressions in
the interpreter are paired up with an environment, i.e., they are
closures.

There are a few places where I think we may still have bugs in our
handling of size closures, but I think they are quite obscure.