ad: fewer bounds checks (#2247)
This changes most of the UpdateAcc operations emitted by the VJP
transform to not have bounds checks. The idea is that if they worked
in the forward sweep, they ought also to be in-bounds in the return
sweep (there are some exceptions).

I don't fully remember why we had so many bounds checks in the first
place. I do know that we had them in the initial design for
commonality with 'scatter', and it is possible that we never went back
and thought about whether they were truly necessary.