wip
I don't like this approach. Maybe having Bijective/FiltPartInv
also add Injective/Injective+Bijective properties is the
way to go?
Also maybe we really do want the Img to be part of
the FiltPartInv property instead of inferring it.
(Would have to check Img unifies with sum of pf
in proof, but we already do this successfully
in the BijectiveRCD proof, I think.)
Finally, alreadyKnown probably has a lot of duplicated
code.