Fix #2258.
I am somewhat uncertain about this approach, which is based on playing fast and
loose with these existential sizes. What happens if the function is applied
twice? It will overwrite the same existential! I suspect a better approach is to
rework how we compute the shapes for sum type constructors. Specifically, we
should look at the payload values, and *only* in case that does not fully
specify the shape should we look at the type annotation. The type system
guarantees that in the event that is needed, the type will not contain any
existentials (or at least those existentials can be resolved by the concrete
payload).