Begin tracking provenance of IR statements.
The goal is to eventually provide more useful profiling information, with
references to the original source code. However, it is crucial that this does
not overly burden the various compiler passes.

The key trick here is to stuff the provenance into the StmAux, which we always
propagate (hopefully) because it is where we stuff attributes and certificates
as well.