interpreter: handle sign/unsign for AD values.
This is really just papering over a bug due to the interpreter's AD
implementation using a different scalar representation that does not track
signedness. However it is enough to make it for functions that only use
signing/unsigning internally, and besides, it is not meaningful to differentiate
functions that return integers.

Also makes some cosmetic changes.