Fix #2292.
This turned out to be a little bit more involved than I thought, as it was
related to improper handling of built-in functions that have different return
and argument types.

The ultimate solution was to change the to_bits/from_bits functions from being
built-in functions to instead being conversion operators - this significantly
simplifies the logic.