Support CUDA 13.
The cuCtxCreate() function apparently grew a new parameter in CUDA 13.
Unfortunately, the CUDA documentation is written very poorly for this new
parameter, so I do not understand exactly what it is for. Internet resources
suggest that it can simply be NULL, but I do not have a CUDA 13 system on which
I can test the veracity of this claim.

Fixes #2327.