quickhull works.
Caveats:
 * loop is commented out but we verify the loop
body by extracting it into a top-level
function annotated with identical preconditions
and postconditions. This way the pre- and postcondition
acts as a loop invariant.
We verify that first call to the loop body
satisfies the preconditions---all subsequent
calls are then given by the loop invariant.
 * I commented out the conditional handling
empty arrays in semihull. Our implementation
doesn't support conditionals that return
different sized arrays yet and the case
is just a shortcircuit to not perform
the semihull computation (there's nothing to check
there).