Fix bug duplicating names in Algebra layer, losing properties.
Predicates and sums (with stride different from 1) are not translatable
to the Algebra layer. Therefore we may have sums over untranslatable
expressions.

For example, the following are currently untranslatable
  * Sum (x+y) --> Sum x + Sum y.
  * Sum_j (x[j] > 0) --> Sum v[j] where fresh v |-> x[hole].
  * Sum_j (x[y[j]] > 0) --> Sum v[j] where fresh v |-> x[y[hole]].
  * Sum_{j=0}^{n} (x[j + 1]) --> Sum v[0:n] where fresh v |-> x[hole + 1].
The name v depends on the context and we try to
create the holes to match the context heuristically.

The latter expression can be made less opaque by instead shifting the
upper and lower boudns of the sum, because it has stride 1.
  * Sum_{j=0}^{n} (x[j + 1]) --> Sum x[1:n+1]
This is beneficial when there are properties recorded on x (e.g., Range,
Monotonicity) that otherwise has to be propagated to the fresh name v.
That process is fragile and complicates the translation.
Also it makes the algebra layer output less opaque when debugging.

Caveat: in the back-translation, the sum has now changed to
Sum_{j=1}^{n+1} (x[j]). I can mitigate this by keeping a table over the
bounds, but first thing is to see whether this caveat even matters.

Limitation: This removes propagation of properties to untranslatable
names to make the code cleaner. Therefore only properties on sums with
(provable) stride 1 enjoy property information.